requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch
torchvision
clip @ git+https://github.com/yonatanbitton/CLIP
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
ftfy==6.1.1
flask_cors
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
regex==2022.4.24
requests==2.27.1
six==1.16.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
waitress